Refactored again to be more like a library. Removed existing examples, replaced with something more simple.
