+++ examples/is_that_what_she_said.py
@@ -1,0 +1,7 @@
+import twss
+
+sentance = "It's too hard"
+print sentance, twss.is_positice(sentance)
+
+sentance = "I like jam"
+print sentance, twss.is_positive(sentance)
