--- templates/inputform.html
@@ -1,11 +1,0 @@
-$def with (form, istwss)
-
-$if istwss is True:<h3>TWSS!</h3>
-
-$if istwss is False:<h3>Not TWSS.</h3>
-
-<form name="main" method="post"> 
-$if not form.valid: <p class="error">Try again</p>
-$:form.render()
-<p><input type="submit" /></p>
-</form>
