+++ tests/twss/test_twss.py
@@ -1,0 +1,28 @@
+import os
+import tempfile
+import unittest
+import twss
+from twss import classifier
+
+class TestTextClassifier(unittest.TestCase):
+
+	def test_is_positive(self):
+		result = twss.is_positive('That was not so hard')
+		self.assertEquals(True, result)
+		
+		result = twss.is_positive('I am a normal sentance')
+		self.assertEquals(False, result)
+	
+	def test_save_and_load(self):
+		temp = tempfile.NamedTemporaryFile(delete=False)
+		temp.close()
+		
+		twss.twss.save(temp.name)
+		classifier = twss.classifier.TextClassifier.load(temp.name)
+		
+		result = classifier.is_positive('That was not so hard')
+		self.assertEquals(True, result)
+		
+		result = classifier.is_positive('I am a normal sentance')
+		self.assertEquals(False, result)
+		os.unlink(temp.name)
