+++ twss/data/default_TextClassifier.pickle
@@ -1,4 +1,4 @@
-(itwsslib
+(iclassifier
 TextClassifier
 p0
 (dp1
